Add and refine full postSQL functionality

+++ controllers/coasterController.js
@@ -1,4 +1,4 @@
-const db = require('../db/queries');
+const db = require('../db/coasterQueries');
 
 async function getAllCoasters(req, res) {
     const coasters = await db.getAllCoasters();
@@ -15,14 +15,24 @@ function getNewCoasterForm(req, res) {
 };
 
 async function postNewCoaster(req, res) {
-    const { coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength } = req.body;
-    await db.postNewCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength);
+    const { coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName } = req.body;
+    await db.postNewCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName);
     res.redirect('/coaster');
 };
 
 async function getUpdateCoasterForm(req, res) {
     const id = Number(req.params.id);
+
+    if (Number.isNaN(id)) {
+        return res.status(400).send('Invalid ID');
+    };
+
     const coaster = await db.findCoasterById(id);
+
+    if (!coaster) {
+        return res.status(404).render('404');
+    };
+
     res.render('updateCoasterForm', {
         title: 'Update Coaster',
         coaster: coaster,
@@ -30,15 +40,30 @@ async function getUpdateCoasterForm(req, res) {
 };
 
 async function postUpdatedCoaster(req, res) {
-    const { coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength } = req.body;
+    const { coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, parkName } = req.body;
     const id = Number(req.params.id);
-    await db.updateExistingCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, id)
+
+    if (Number.isNaN(id)) {
+        return res.status(400).send('Invalid ID');
+    };
+
+    await db.updateExistingCoaster(coasterName, coasterInversions, coasterSpeed, coasterHeight, coasterLength, id, parkName);
     res.redirect('/coaster');
 };
 
 async function getSingleCoaster(req, res) {
     const id = Number(req.params.id);
+
+    if (Number.isNaN(id)) {
+        return res.status(400).send('Invalid ID');
+    };
+
     const coaster = await db.findCoasterById(id);
+
+    if (!coaster) {
+        return res.status(404).render('404');
+    };
+
     res.render('coasterData', {
         title: 'Coaster Data',
         coaster: coaster,
@@ -47,8 +72,50 @@ async function getSingleCoaster(req, res) {
 
 async function deleteSingleCoaster(req, res) {
     const id = Number(req.params.id);
+
+    if (Number.isNaN(id)) {
+        return res.status(400).send('Invalid ID');
+    };
+
     await db.deleteCoasterById(id);
     res.redirect('/coaster');
+};
+
+async function getAddToRiderForm(req, res) {
+    const id = Number(req.params.id);
+
+    if (Number.isNaN(id)) {
+        return res.status(400).send('Invalid ID');
+    };
+
+    const coaster = await db.findCoasterById(id);
+
+    if (!coaster) {
+        return res.status(404).render('404');
+    };
+
+    res.render('addRiderFromCoasterForm', {
+        title: 'Add Coaster to Rider',
+        coaster: coaster,
+    });
+};
+
+async function postCoasterToRider(req, res) {
+    const coasterId = Number(req.params.id);
+
+    if (Number.isNaN(coasterId)) {
+        return res.status(400).send('Invalid ID');
+    };
+
+    const { riderName } = req.body;
+    const riderId = await db.getRiderIdFromName(riderName);
+    
+    if(!riderId) {
+        return res.status(404).send('Rider not found');
+    };
+
+    await db.addCoasterToRiderById(coasterId, riderId);
+    res.redirect(`/coaster/${coasterId}`);
 };
 
 module.exports = {
@@ -59,4 +126,6 @@ module.exports = {
     deleteSingleCoaster,
     getUpdateCoasterForm,
     postUpdatedCoaster,
+    getAddToRiderForm,
+    postCoasterToRider,
 };
